refactor(tui): split TUI helpers by concern (Phase 11)
Implements Phase 11 of the Rust structure refactor roadmap. Splits
`src/tui.rs` (774 LOC) into three concern-focused files plus a thin
entrypoint `mod.rs`.

Split (actual usage-driven, not just plan-prescribed):

- `mod.rs` (30 LOC, was 774) — `DEBUG_MODE` atomic + `set_debug_mode`,
  shared color constants (`WHITE`, `PHOSPHOR_GREEN`, `PHOSPHOR_DIM`,
  `PHOSPHOR_DARK`), the `rgb` helper, module declarations, and
  `pub use` re-exports for every previously-public function.
- `animation.rs` (502 LOC) — intro/outro/digital-rain code
  (`intro_animation`, `outro_animation`, `simple_outro`,
  `digital_rain`, `type_text`, `glitch_text`, the `RainCell` struct,
  `skippable_sleep`, `DIM` color, all the rain randomness helpers).
- `output.rs` (176 LOC) — tables, hints, fatal output, logo printing,
  terminal title (`print_config_table`, `step_shimmer`, `step_quiet`,
  `step_fail`, `print_deploying`, `print_logo`, `fatal`,
  `set_terminal_title`, `clear_screen`, `shorten_home`, `hint`, plus
  the `ROSE` color used only here).
- `prompt.rs` (95 LOC) — `prompt_choice` and the generic `spin_wait`.

Verbatim moves. No color edits, no timing tweaks, no render changes.

Shared-helper placement decisions (usage-verified before moving):
- `WHITE` is used by both animation and output → stays in `mod.rs`.
- `PHOSPHOR_GREEN/DIM/DARK` used across all three submodules → stays
  in `mod.rs` (plan listed these under animation, but empirical
  usage is broader; the plan allows inspection-before-moving).
- `rgb` helper used across all three → stays in `mod.rs`.
- `DIM` only used by `animation::simple_outro` → moved to animation.
- `ROSE` only used by output's `step_fail`/`fatal` → moved to output.
- `skippable_sleep` only used by animation → moved to animation.
- `DEBUG_MODE` stays in `mod.rs`; child modules access it via
  `super::DEBUG_MODE` without a visibility bump (Rust child modules
  can read parent-private items).

No tests to migrate: `src/tui.rs` had no `#[cfg(test)] mod tests`
block.

Visibility: zero bumps required. Every previously-`pub` function
remains `pub` and is re-exported from `mod.rs`. Shared helpers in
`mod.rs` stay module-private and are reached via `super::` from
children.

Zero call sites outside `src/tui/` required edits. All 35+ external
references (across `src/main.rs`, `src/app/`, `src/launch.rs`,
`src/runtime/`) resolve unchanged through the re-exports in `mod.rs`.

Verification per COMMITS.md:
- cargo fmt -- --check — pass
- cargo clippy — zero warnings
- cargo nextest run — 418 tests run: 418 passed, 0 skipped (unchanged)
